draft: scala event subscriptions

+++ src/rpc/server/scala/src/main/scala/com/ctc/omega_edit/grpc/EditorService.scala
@@ -93,7 +93,7 @@ class EditorService(implicit val system: ActorSystem) extends Editor {
     case Viewport.Id(sid, vid) =>
       (editors ? ViewportOp(sid, vid, Viewport.Get)).mapTo[Result].map {
         case Err(c)           => throw grpcFailure(c)
-        case ok: Ok with Data => ViewportDataResponse(ok.id, ok.data.size(), ok.data)
+        case ok: Ok with Data => ViewportDataResponse(ok.id, ok.data.size.toLong, ok.data)
         case Ok(id)           => ViewportDataResponse(id)
       }
     case _ => grpcFailFut(Status.INVALID_ARGUMENT, "malformed viewport id")
